+++ src/features/commands.ts
@@ -4,8 +4,63 @@ import * as vscode from 'vscode';
 import * as fs from 'fs';
 import { LanguageClient, ExecuteCommandParams, ExecuteCommandRequest } from 'vscode-languageclient/node';
 import { workspace, window } from 'vscode';
+import { WURST_HOME } from '../paths';
+import { appendDiagnostic, buildDiagnosticsText, formatDiagnosticError } from './diagnostics';
+
+let diagnosticsClient: LanguageClient | null = null;
+
+function showLanguageServerOutput(): void {
+    try {
+        if (diagnosticsClient) {
+            diagnosticsClient.outputChannel.show();
+            return;
+        }
+        void vscode.commands.executeCommand('workbench.action.output.toggleOutput');
+    } catch (error) {
+        appendDiagnostic('VS Code extension', `Could not show language server output: ${formatDiagnosticError(error)}`);
+        void vscode.commands.executeCommand('workbench.action.output.toggleOutput');
+    }
+}
+
+async function copyDiagnostics(): Promise<void> {
+    try {
+        await vscode.env.clipboard.writeText(buildDiagnosticsText(WURST_HOME));
+        void vscode.window.showInformationMessage('Copied Wurst diagnostics to the clipboard.');
+    } catch (error) {
+        const detail = formatDiagnosticError(error);
+        appendDiagnostic('VS Code extension', `Could not copy diagnostics: ${detail}`);
+        void vscode.window.showErrorMessage(`Could not copy Wurst diagnostics: ${detail}`);
+    }
+}
+
+async function openWurstHome(): Promise<void> {
+    try {
+        await vscode.commands.executeCommand('revealFileInOS', vscode.Uri.file(WURST_HOME));
+    } catch (error) {
+        const detail = formatDiagnosticError(error);
+        appendDiagnostic('VS Code extension', `Could not open Wurst home: ${detail}`);
+        void vscode.window.showErrorMessage(`Could not open Wurst home: ${detail}`);
+    }
+}
+
+export function registerWurstDiagnosticsCommands(context: vscode.ExtensionContext): void {
+    context.subscriptions.push(
+        vscode.commands.registerCommand('wurst.openWurstHome', () => openWurstHome()),
+        vscode.commands.registerCommand('wurst.copyDiagnostics', () => copyDiagnostics()),
+        vscode.commands.registerCommand('wurst.showLogs', () => showLanguageServerOutput()),
+        vscode.commands.registerCommand('wurst.showDiagnosticsActions', async () => {
+            const choice = await vscode.window.showQuickPick([
+                { label: '$(folder-opened) Open Wurst home', command: 'wurst.openWurstHome' },
+                { label: '$(copy) Copy diagnostics', command: 'wurst.copyDiagnostics' },
+                { label: '$(output) Open Wurst output', command: 'wurst.showLogs' },
+            ], { placeHolder: 'Wurst diagnostics' });
+            if (choice) await vscode.commands.executeCommand(choice.command);
+        }),
+    );
+}
 
 export function registerCommands(client: LanguageClient): vscode.Disposable {
+    diagnosticsClient = client;
     let _lastMapConfig: string | undefined = undefined;
 
     // Accepts both archive files (*.w3x, *.w3m) and folder-mode directories (*.w3x/, *.w3m/)
